able to get comments on DOM, need to debug author.id error
