--- pages/comments/index.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function ViewComments() {
-  return (
-    <div>index</div>
-  );
-}
